savedjob, mockinterview(finish), update UI

+++ FlexiConnectBE/src/main/java/com/soodthin/dto/response/AI/N8nAnalysisResponse.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @Data
 public class N8nAnalysisResponse {
     @JsonProperty("ai_score")
-    private Integer aiScore;
+    private Double aiScore;
     
     @JsonProperty("feedback")
     private String feedback;
